+++ lib/api/entity.dart
@@ -1,11 +1,13 @@
 import "dart:async";
 import 'dart:developer';
 import 'dart:typed_data';
+import 'package:dvote/blockchain/ens.dart';
 import 'package:dvote/net/gateway-pool.dart';
 import 'package:dvote/net/gateway-web3.dart';
 import 'package:dvote/wrappers/content-uri.dart';
 import 'package:convert/convert.dart';
 import 'package:dvote/util/parsers.dart';
+import 'package:hex/hex.dart';
 import 'package:dvote/wrappers/entities.dart';
 
 import '../models/build/dart/metadata/entity.pb.dart';
@@ -22,18 +24,22 @@ Future<EntityMetadata> fetchEntity(
 
   // Fetch the Content URI from the blockchain
   final hexEntityId = hex.decode(entityRef.entityId.substring(2));
+  print(entityRef.entityId);
+  print(HEX.encode(ensHashAddress(Uint8List.fromList(hexEntityId))));
   try {
     final params = [
-      Uint8List.fromList(hexEntityId),
+      ensHashAddress(Uint8List.fromList(hexEntityId)),
       TextRecordKeys.JSON_METADATA_CONTENT_URI
     ];
+    print(params);
     result = await gw.callMethod("text", params, ContractEnum.EntityResolver);
+    print(result);
     if (result == null || result.length == 0 || result.first == null)
       throw Exception("The metadata of the entity can't be found");
     else if (result[0] is! String || result[0].length == 0)
       throw Exception("The response from the blockchain is invalid");
-  } catch (err) {
-    log(err.toString());
+  } catch (err, s) {
+    log("${err.toString()}, $s");
     throw err;
   }